requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.3
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-rename-app==0.1.5
pycodestyle==2.10.0
sqlparse==0.4.3
tomli==2.0.1
